refactor(repository): repository.payment 파일을 repository로 이동
6개의 repository 파일을 `repository.payment` 패키지에서 `repository` 패키지로 이동하여 패키지 구조를 단순화시킴.
